Move wrapper under pkg/
