+++ internal/language/language.go
@@ -13,7 +13,7 @@ import (
 	"github.com/anttiharju/vmatch/internal/exitcode"
 	"github.com/anttiharju/vmatch/internal/find"
 	"github.com/anttiharju/vmatch/internal/install"
-	"github.com/anttiharju/vmatch/internal/wrapper"
+	"github.com/anttiharju/vmatch/pkg/wrapper"
 )
 
 type WrappedLanguage struct {
